+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 ## Changelog
 
+**1.1.0 Changes:**
+
+* Pressure plates on Abandoned Aqueduct can now be stolen
+* The Timed Chest on Rallypoint Delta can now be stolen
+* Fans on Rallypoint Delta can now be stolen
+
 **1.0.0**
 
 * First release